+++ CHANGELOG.md
@@ -21,6 +21,8 @@ new minor version numbers. Please note that this is no version handling covered 
 
 - Add support of lots of unicode named classes like `XID_Start` and `XID_Continue` by the help of
 the `seshat-unicode` crate
+- Performance: Scanner holds ScannerImpl in a `Rc<RefCell<>>` to save time during creation of a new
+`find_iter`
 
 ## 0.3.0 - 2024-08-29